implement ruff-native cycle detection, code reorg
