+++ reference/README.md
@@ -1,0 +1,4 @@
+The code in this folder is just for reference, to help play with logic
+before implementing in Rust. There is no guarantee that is is fully
+in sync with the canonical implementation, or even that it is runnable
+as a standard tool
