+++ src/main.rs
@@ -1,30 +1,35 @@
-mod minimal_cycles;
+mod cycle_detection;
+mod minimize_cycles;
+mod ruff_util;
 
 use structopt::StructOpt;
 
 #[derive(StructOpt, Debug)]
 pub struct MinimizeCyclesOptions {
     #[structopt(short, long)]
-    repo_root: String,
-
-    #[structopt(short, long)]
     cycle_results_file: String,
 }
 
 #[derive(StructOpt, Debug)]
+pub struct CycleDetectionOptions {}
+
+#[derive(StructOpt, Debug)]
 enum RuffTools {
+    #[structopt(name = "detect-cycles")]
+    CycleDetection(CycleDetectionOptions),
     #[structopt(name = "minimize-cycles")]
     MinimizeCycles(MinimizeCyclesOptions),
+    #[structopt(name = "lint-imports")]
     ImportLinter {},
-    CycleDetection {},
+    #[structopt(name = "live")]
+    Live {},
 }
 
 fn main() {
     let options = RuffTools::from_args();
     match options {
-        RuffTools::MinimizeCycles(cmd) => {
-            minimal_cycles::minimize_cycles(cmd.repo_root, cmd.cycle_results_file)
-        }
+        RuffTools::MinimizeCycles(cmd) => minimize_cycles::minimize_cycles(cmd.cycle_results_file),
+        RuffTools::CycleDetection(_) => cycle_detection::detect_cycles(),
         _ => (),
     }
 }
